pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "fuzzywuzzy>=0.18.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "python-levenshtein>=0.27.1",
    "requests>=2.32.4",
]
